Trying float16

+++ lowdim.py
@@ -63,5 +63,5 @@ savemat(outputFile+"/"+"evals.mat",mdict={'evals':latent},oned_as='column',do_co
 for ik in range(0,k):
 	logging.info("(lowdim) writing scores for pc " + str(ik))
 	print(str(shape(inner(resp.first()-meanVec,sortedDim2[ik,:]))))
-	out = resp.map(lambda x : inner(x-meanVec,sortedDim2[ik,:]))
+	out = resp.map(lambda x : float16(inner(x-meanVec,sortedDim2[ik,:])))
 	savemat(outputFile+"/"+"scores-"+str(ik)+".mat",mdict={'scores':out.collect()},oned_as='column',do_compression='true')
